Task controller in process
